requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
datasets==2.20.0
dill==0.3.8
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.24.5
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
Rtree==1.3.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.66.5
trimesh==4.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
GPUtil
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
